vanilla parity
passes a set of tests that compare frameCounter's low byte, both
rng_bytes and the gameMode byte with vanilla's at the same time.

current state:
game modes 0 1 and 2 are all defined in branch.asm
menu is not reachable
first 255 frames of legal screen can't be skipped

intentional differences:
vanilla switches to gameMode 6 (demo setup) (5 by the the the next
frame comes around) when gameMode is 1 (title screen) and the high
byte of the frameCounter is 5.  This switches to gameMode 2 (game type
menu) instead

When doing nothing, this happens at frame 1795
