+++ src/modes/crash.asm
@@ -221,7 +221,7 @@ testCrash:
         inc allegroIndex
 @newBit0:
         lda nmiReturnAddr
-        cmp #<updateAudioWaitForNmiAndResetOamStaging+10
+        cmp #<(updateAudioWaitForNmiAndResetOamStaging+10)
         beq @returnLate ; checking which instruction returned to. if so, add 3 cycles
         lda #$03
         clc
