+++ tests/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 gumdrop = "0.8.1"
 rustico-core = { git = "https://github.com/zeta0134/rustico.git", rev = "d31d380b1e15bf7cc80e8827730e56a6df77ae46" }
 minifb = "0.27"
+gag = "1.0.0"
 
 # for patch testing
 flips = "0.2.1"
